add realpayload test, and fix bug during payload parse
