+++ src/test/resources/com.uid2.shared/test/secure/gcpoaid/jwt_payload_policy_valid.json
@@ -1,0 +1,57 @@
+{
+  "aud": "https://sts.googleapis.com",
+  "exp": 1688132564,
+  "iat": 1688128964,
+  "iss": "https://confidentialcomputing.googleapis.com",
+  "nbf": 1688128964,
+  "sub": "https://www.googleapis.com/compute/v1/projects/someone-secundus-bank/zones/us-west1-b/instances/secundus-initial-vm",
+  "secboot": true,
+  "oemid": 11129,
+  "hwmodel": "GCP_AMD_SEV",
+  "swname": "CONFIDENTIAL_SPACE",
+  "swversion": [
+    "230600"
+  ],
+  "dbgstat": "disabled-since-boot",
+  "submods": {
+    "confidential_space": {
+      "support_attributes": [
+        "LATEST",
+        "STABLE",
+        "USABLE"
+      ]
+    },
+    "container": {
+      "image_reference": "us-docker.pkg.dev/someone-primus-bank/primus-workloads/initial-workload-container:latest",
+      "image_digest": "sha256:fedbd6eaf93394d5eb98d7e52d4cce17e0ea23f7fac1a6bd68e097ca85a4fffb",
+      "restart_policy": "Never",
+      "image_id": "sha256:5be33a19451733a45ea1bdb340fcb858a0fc733e91ca0a0d99638652f6dcabd0",
+      "env_override": {
+        "env": "prod",
+        "scope": "uid2",
+        "api_key": "dummy"
+      },
+      "cmd_override": null,
+      "env": {
+        "HOSTNAME": "secundus-initial-vm",
+        "PATH": "/usr/local/sbin:/usr/local/bin:/usr/sbin:/usr/bin:/sbin:/bin"
+      },
+      "args": [
+        "/test/initial_workload",
+        "count-location",
+        "Seattle",
+        "gs://someone-secundus-bank-results-storage/seattle-result"
+      ]
+    },
+    "gce": {
+      "zone": "us-west1-b",
+      "project_id": "someone-secundus-bank",
+      "project_number": "600005144714",
+      "instance_name": "secundus-initial-vm",
+      "instance_id": "5180554646064952245"
+    }
+  },
+  "google_service_accounts": [
+    "run-confidential-vm@someone-secundus-bank.iam.gserviceaccount.com"
+  ]
+}
